Factor out validation entries into own module.
